+++ config.py
@@ -11,7 +11,7 @@ Auteur        : Pierre Théberge
 Compagnie     : Innovations, Performances, Technologies inc.
 Créé le       : 2025-08-05
 Modifié le    : 2026-09-23
-Version       : 0.5.23
+Version       : 0.5.24
 Copyright     : Pierre Théberge
 
 Description
@@ -114,6 +114,7 @@ Modifications
 0.5.21 - 2026-08-18   ES-34   : Synchronisation de version (aucun changement fonctionnel).
 0.5.22 - 2026-08-20   ES-34   : Synchronisation de version (aucun changement fonctionnel).
 0.5.23 - 2026-09-23   ES-28   : Synchronisation de version (aucun changement fonctionnel).
+0.5.24 - 2026-09-23   CR      : Synchronisation de version (aucun changement fonctionnel).
 
 Paramètres
 ----------